medbrains/qa/requirements.txt
tcms-api>=13.0
pyyaml>=6.0
python-dotenv>=1.0
